requirements.txt
Django==1.6
Pillow==2.2.1
South==0.8.4
argparse==1.2.1
django-tastypie==0.10.0
mimeparse==0.1.3
psycopg2==2.5.2
python-dateutil==2.2
six==1.4.1
wsgiref==0.1.2
